requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
django-autoslug==1.9.6
django-crispy-forms==1.8.1
django-extensions==2.2.6
django-froala-editor==3.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
pyTelegramBotAPI==3.6.7
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
